Refactor and enhance testing framework: - Update VSCode settings for Python imports and spell checking. - Revise pytest commands in tasks.json for better test coverage reporting. - Add new test cases for CUDA device handling and model loading. - Introduce new IO utility functions for file and image handling. - Clean up utility functions and remove redundant code. - Implement tracing functionality for module imports.
