+++ tests/test_util.py
@@ -1,25 +1,8 @@
-import os
-import tempfile
-
 import numpy as np
 import pytest
 import torch
 
-from vut.util import (
-    get_dirs,
-    init_seed,
-    load_list,
-    load_np,
-    load_tensor,
-    save,
-    save_list,
-    save_np,
-    save_tensor,
-    to_list,
-    to_np,
-    to_tensor,
-    unique,
-)
+from vut.util import init_seed, to_list, to_np, to_tensor, unique
 
 
 def test_init_seed__same_values():
@@ -93,31 +76,6 @@ def test_unique__unsupported_type():
         unique("unsupported type")
 
 
-def test_get_dirs__non_recursive():
-    path = os.path.dirname(__file__)
-    dirs = get_dirs(path, recursive=False)
-    assert all(os.path.isdir(d) for d in dirs), "All items should be directories"
-
-
-def test_get_dirs__recursive():
-    path = os.path.dirname(__file__)
-    dirs = get_dirs(path, recursive=True)
-    assert all(os.path.isdir(d) for d in dirs), "All items should be directories"
-
-
-def test_get_dirs__empty_directory():
-    with tempfile.TemporaryDirectory() as temp_dir:
-        dirs = get_dirs(temp_dir, recursive=False)
-        assert dirs == [], "Empty directory should return an empty list"
-
-
-def test_get_dirs__non_existent_path():
-    with tempfile.TemporaryDirectory() as temp_dir:
-        non_existent_path = os.path.join(temp_dir, "non_existent")
-        dirs = get_dirs(non_existent_path, recursive=False)
-        assert dirs == [], "Non-existent path should return an empty list"
-
-
 def test_to_list__list():
     data = [1, 2, 3]
     result = to_list(data)
@@ -185,135 +143,3 @@ def test_to_tensor__tensor():
 def test_to_tensor__unsupported_type():
     with pytest.raises(TypeError):
         to_tensor("unsupported type")
-
-
-def test_save_list():
-    data = [1, 2, 3]
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        file_path = temp_file.name
-        save_list(data, file_path)
-    with open(file_path, "r") as f:
-        content = f.read()
-    assert content == "1\n2\n3\n", "File content should match the list"
-    os.remove(file_path)
-
-
-def test_save_list_with_callback():
-    data = [1, 2, 3]
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        file_path = temp_file.name
-        save_list(data, file_path, callback=lambda x: f"{x}0")
-    with open(file_path, "r") as f:
-        content = f.read()
-    assert content == "10\n20\n30\n", "File content should match the list"
-    os.remove(file_path)
-
-
-def test_save_np():
-    data = np.array([1, 2, 3])
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        file_path = temp_file.name + ".npy"
-        save_np(data, file_path)
-    loaded_data = np.load(file_path)
-    assert np.array_equal(loaded_data, data), (
-        "Loaded data should match the original array"
-    )
-    os.remove(file_path)
-
-
-def test_save_tensor():
-    data = torch.tensor([1, 2, 3])
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        file_path = temp_file.name
-        save_tensor(data, file_path)
-    loaded_data = torch.load(file_path)
-    assert torch.equal(loaded_data, data), (
-        "Loaded data should match the original tensor"
-    )
-    os.remove(file_path)
-
-
-def test_save__list():
-    data = [1, 2, 3]
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        file_path = temp_file.name
-        save(data, file_path)
-    with open(file_path, "r") as f:
-        content = f.read()
-    assert content == "1\n2\n3\n", "File content should match the list"
-    os.remove(file_path)
-
-
-def test_save__ndarray():
-    data = np.array([1, 2, 3])
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        file_path = temp_file.name + ".npy"
-        save(data, file_path)
-    loaded_data = np.load(file_path)
-    assert np.array_equal(loaded_data, data), (
-        "Loaded data should match the original array"
-    )
-
-
-def test_save__tensor():
-    data = torch.tensor([1, 2, 3])
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        file_path = temp_file.name
-        save(data, file_path)
-    loaded_data = torch.load(file_path)
-    assert torch.equal(loaded_data, data), (
-        "Loaded data should match the original tensor"
-    )
-
-
-def test_save__unsupported_type():
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        file_path = temp_file.name
-        with pytest.raises(TypeError):
-            save("unsupported type", file_path)
-
-
-def test_load_list():
-    data = [1, 2, 3]
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False) as temp_file:
-        file_path = temp_file.name
-        temp_file.writelines(f"{item}\n" for item in data)
-    loaded_data = load_list(file_path)
-    assert loaded_data == [str(i) for i in data], (
-        "Loaded data should match the original list"
-    )
-    os.remove(file_path)
-
-
-def test_load_list_with_callback():
-    data = [1, 2, 3]
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False) as temp_file:
-        file_path = temp_file.name
-        temp_file.writelines(f"{item}\n" for item in data)
-    loaded_data = load_list(file_path, callback=int)
-    assert loaded_data == data, "Loaded data should match the original list"
-    os.remove(file_path)
-
-
-def test_load_np():
-    data = np.array([1, 2, 3])
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        file_path = temp_file.name + ".npy"
-        np.save(file_path, data)
-    loaded_data = load_np(file_path)
-    assert np.array_equal(loaded_data, data), (
-        "Loaded data should match the original array"
-    )
-    os.remove(file_path)
-
-
-def test_load_tensor():
-    data = torch.tensor([1, 2, 3])
-    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-        file_path = temp_file.name
-        torch.save(data, file_path)
-    loaded_data = load_tensor(file_path)
-    assert torch.equal(loaded_data, data), (
-        "Loaded data should match the original tensor"
-    )
-    os.remove(file_path)
