+++ vut/__init__.py
@@ -1,9 +1,10 @@
 from .cuda import get_device, load_model, save_model
-from .mapping import to_class_index, to_class_name
-from .time import format_time, get_elapsed_time, get_time
-from .util import (
+from .io import (
     get_dirs,
-    init_seed,
+    get_images,
+    load_file,
+    load_image,
+    load_images,
     load_list,
     load_np,
     load_tensor,
@@ -11,21 +12,27 @@ from .util import (
     save_list,
     save_np,
     save_tensor,
-    to_list,
-    to_np,
-    to_tensor,
-    unique,
 )
+from .mapping import to_class_index, to_class_name
+from .time import format_time, get_elapsed_time, get_time
+from .trace import disable_trace, enable_trace
+from .util import init_seed, to_list, to_np, to_tensor, unique
 
 __all__ = [
+    "disable_trace",
+    "enable_trace",
     "format_time",
     "init_seed",
+    "load_file",
+    "load_image",
+    "load_images",
     "load_list",
     "load_model",
     "load_np",
     "load_tensor",
     "get_device",
     "get_dirs",
+    "get_images",
     "get_time",
     "get_elapsed_time",
     "save",
