+++ vut/trace.py
@@ -1,7 +1,19 @@
 import sys
+from types import FrameType
+from typing import Any
 
 
-def trace_module(frame, event: str, arg):
+def trace_module(frame: FrameType, event: str, arg: Any):
+    """Trace module imports.
+
+    Args:
+        frame (FrameType): The current stack frame.
+        event (str): The event type (e.g., "call", "return").
+        arg (Any): Additional event-specific information.
+
+    Returns:
+        Any: The return value of the traced function or None.
+    """
     code = frame.f_code
     module_name = code.co_filename
     if (
@@ -15,8 +27,10 @@ def trace_module(frame, event: str, arg):
 
 
 def enable_trace():
+    """Enable tracing of module imports."""
     sys.settrace(trace_module)
 
 
 def disable_trace():
+    """Disable tracing of module imports."""
     sys.settrace(None)
